\name{aws-package}
\alias{aws-package}
\docType{package}
\title{
\packageTitle{aws}
}
\description{
\packageDescription{aws}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{aws}
\packageIndices{aws}

}
\author{
\packageAuthor{aws}

Maintainer: \packageMaintainer{aws}
}
\references{ J. Polzehl and V. Spokoiny (2006)
             Propagation-Separation Approach for Local Likelihood Estimation,
             \emph{Prob. Theory and Rel. Fields} \bold{135(3)}, 335-362.

             J. Polzehl and V. Spokoiny (2004)
             Spatially adaptive regression estimation: Propagation-separation approach,
             WIAS-Preprint 998.

             V. Katkovnik, K. Egiazarian and J. Astola (2006)
             \emph{Local Approximation Techniques in Signal and Image Processing},
              SPIE Press Monograph Vol. PM 157
}
\keyword{ package }
