\name{aws-package}
\alias{aws-package}
\docType{package}
\title{
\packageTitle{aws}
}
\description{
\packageDescription{aws}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{aws}
\packageIndices{aws}

}
\author{
\packageAuthor{aws}

Maintainer: \packageMaintainer{aws}
}
\references{ J. Polzehl, K. Tabelow (2019). Magnetic Resonance Brain Imaging:
             Modeling and Data Analysis Using R. Springer, Use R! series. Appendix A.
             Doi:10.1007/978-3-030-29184-6.

             J. Polzehl, K. Papafitsoros, K. Tabelow. Patch-wise adaptive weights smoothing,
             Preprint no. 2520, WIAS, Berlin, 2018, DOI 10.20347/WIAS.PREPRINT.2520.
             (to appear in Journal of Statistical Software).

             J. Polzehl and V. Spokoiny (2006)
             Propagation-Separation Approach for Local Likelihood Estimation,
             Prob. Theory and Rel. Fields 135(3), 335-362. DOI:10.1007/s00440-005-0464-1.

             J. Polzehl, V. Spokoiny,  Adaptive Weights Smoothing with applications to image restoration,
						 J. R. Stat. Soc. Ser. B Stat. Methodol.  62 ,  (2000)  , pp. 335--354. DOI:10.1111/1467-9868.00235.

             V. Katkovnik, K. Egiazarian and J. Astola (2006)
             Local Approximation Techniques in Signal and Image Processing,
              SPIE Press Monograph Vol. PM 157

							A. Buades, B. Coll and J. M. Morel (2006). A review of image denoising algorithms,
							with a new one.
							Simulation, 4, 490-530. DOI:10.1137/040616024.

							Rudin, L.I., Osher, S. and Fatemi, E. (1992). Nonlinear total variation based
					    noise removal algorithms. Phys. D, 60, 259-268. DOI: 10.1016/0167-2789(92)90242-F.

					    Bredies, K., Kunisch, K. and Pock, T. (2010). Total Generalized Variation.
					    SIAM J. Imaging Sci., 3, 492-526. DOI:10.1137/090769521.

}
\keyword{ package }
